+++ apps/macos/Sources/OpenClawMacCLI/WizardCommand.swift
@@ -273,7 +273,7 @@ actor GatewayWizardClient {
         }
         let connectNonce = try await self.waitForConnectChallenge()
         let identity = DeviceIdentityStore.loadOrCreate()
-        let signedAtMs = Int(Date().timeIntervalSince1970 * 1000)
+        let signedAtMs = Int64(Date().timeIntervalSince1970 * 1000)
         let payload = GatewayDeviceAuthPayload.buildConnectCompatibilityPayload(
             deviceId: identity.deviceId,
             clientId: clientId,
