+++ apps/shared/OpenClawKit/Sources/OpenClawKit/GatewayChannel.swift
@@ -564,7 +564,7 @@ public actor GatewayChannelActor {
         } else if let password = selectedAuth.authPassword {
             params["auth"] = ProtoAnyCodable(["password": ProtoAnyCodable(password)])
         }
-        let signedAtMs = Int(Date().timeIntervalSince1970 * 1000)
+        let signedAtMs = Int64(Date().timeIntervalSince1970 * 1000)
         let connectNonce = try await self.waitForConnectChallenge()
         if includeDeviceIdentity, let identity {
             let payload = GatewayDeviceAuthPayload.buildConnectCompatibilityPayload(
